requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.3.35
sqlmodel>=0.0.14
yfinance>=0.2.32
ddgs>=9.0.0
apscheduler>=3.10.4
openai>=1.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.20.0

# New dependencies for enterprise refactoring
alembic>=1.13.0
pydantic-settings>=2.1.0
tenacity>=8.2.0
